Penambahan API

+++ app/src/main/java/com/example/cobafragment/model/movies/MovieDiscoverResultsItem.java
@@ -36,7 +36,7 @@ public class MovieDiscoverResultsItem {
 	private double popularity;
 
 	@SerializedName("vote_average")
-	private int voteAverage;
+	private double voteAverage;
 
 	@SerializedName("id")
 	private int id;
@@ -127,11 +127,11 @@ public class MovieDiscoverResultsItem {
 		return popularity;
 	}
 
-	public void setVoteAverage(int voteAverage){
+	public void setVoteAverage(double voteAverage){
 		this.voteAverage = voteAverage;
 	}
 
-	public int getVoteAverage(){
+	public double getVoteAverage(){
 		return voteAverage;
 	}
